backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
requests==2.32.3
setuptools==79.0.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
